Suppress stderr
If there are no sessions, tmux will write a message about no server
running to stderr. This breaks the completion script and is totally
irrelevant to the user. So we'll just ignore the stderr messages.

The optimal solution would probably be to extract the tmux call and save
output from it in a variable, test if the exit code from it is 0, and if
it's not, just set COMPREPLY to ().